requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.12.7
cftime==1.6.2
chardet==3.0.4
click==8.1.3
colorama==0.4.6
cycler==0.11.0
dash==2.10.2
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-datetimepicker==0.0.7
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.5
idna==2.10
importlib-metadata==6.6.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kaleido==0.2.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.1.3
netCDF4==1.5.3
numpy==1.24.4
packaging==22.0
pandas==2.0.2
patsy==0.5.3
plotly==5.15.0
pluggy==0.13.1
py==1.11.0
pyparsing==3.0.9
pytest==6.2.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==5.3.1
requests==2.24.0
scipy==1.10.1
six==1.16.0
statsmodels==0.14.0
tenacity==8.2.2
toml==0.10.2
tzdata==2023.3
urllib3==1.25.11
Werkzeug==2.2.3
xarray==0.16.0
zipp==3.15.0
